Move load_location_info_from_file with related class and function to common.py
